requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
cssselect2==0.8.0
daphne==4.2.0
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-csp==4.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
